float to double

+++ promptd0/myAna.h
@@ -5,9 +5,9 @@ namespace ana{
    const float massKaon = 0.493677;
 
    const int nuOfY = 12;
-   const float ybin[nuOfY+1] = {-3, -2.5, -2., -1.5, -1., -0.5, 0, 0.5, 1.0, 1.5, 2.0, 2.5, 3.0};
+   const double ybin[nuOfY+1] = {-3, -2.5, -2., -1.5, -1., -0.5, 0, 0.5, 1.0, 1.5, 2.0, 2.5, 3.0};
    const int nuOfPt = 10;
-   const float ptbin[nuOfPt+1] = {0., 0.5, 1.0, 1.5, 2.0, 3.0, 4.0, 5.0, 6.0, 8.0, 10.0};
+   const double ptbin[nuOfPt+1] = {0., 0.5, 1.0, 1.5, 2.0, 3.0, 4.0, 5.0, 6.0, 8.0, 10.0};
 
    //const double evts_bkg_MB = 1273;
    //const double evts_bkg_central = 121;
